Update: implement new RaspIotRenderer base class and use new audio driver

+++ backend/audio.py
@@ -4,19 +4,22 @@
 import time
 import logging
 import os
-from raspiot.raspiot import RaspIotResource
+from raspiot.raspiot import RaspIotResources
+from raspiot.utils import CommandError, CommandInfo, InvalidParameter, MissingParameter
 from raspiot.libs.commands.alsa import Alsa
 from raspiot.libs.configs.asoundrc import Asoundrc
+from raspiot.libs.drivers.audiodriver import AudioDriver
+from raspiot.libs.drivers.driver import Driver
 
 __all__ = ['Audio']
 
 
-class Audio(RaspIotResource):
+class Audio(RaspIotResources):
     """
     Audio module is in charge of configuring audio on raspberry pi
     """
     MODULE_AUTHOR = u'Cleep'
-    MODULE_VERSION = u'1.0.1'
+    MODULE_VERSION = u'1.1.0'
     MODULE_CATEGORY = u'APPLICATION'
     MODULE_PRICE = 0
     MODULE_DEPS = []
@@ -38,9 +41,37 @@ class Audio(RaspIotResource):
         u'card': 0,
         u'device': 0
     }
-    RESOURCES = {
-        u'audio.capture': 50.0,
-        u'audio.playback': 75.0
+
+    MODULE_RESOURCES = {
+        u'Raspberrypi audio (jack)': {
+            u'audio.playback': {
+                u'hardware_id': u'bcm2835 ALSA',
+                u'permanent': False,
+            }
+        },
+        u'Raspberrypi audio (HDMI)': {
+            u'audio.playback': {
+                u'hardware_id': u'bcm2835 IEC958/HDMI',
+                u'permanent': False,
+            }
+        }
+    }
+
+    AUDIO_DRIVERS = {
+        u'bcm2835 ALSA': {
+            u'output_type': AudioDriver.OUTPUT_TYPE_JACK,
+            u'playback_volume': u'PCM',
+            u'playback_volume_data': (u'Mono', r'\[(\d*)%\]'),
+            u'capture_volume': None,
+            u'capture_volume_data': None,
+        },
+        u'bcm2835 IEC958/HDMI': {
+            u'output_type': AudioDriver.OUTPUT_TYPE_HDMI,
+            u'playback_volume': u'PCM',
+            u'playback_volume_data': (u'Mono', r'\[(\d*)%\]'),
+            u'capture_volume': None,
+            u'capture_volume_data': None,
+        }
     }
 
     def __init__(self, bootstrap, debug_enabled):
@@ -52,10 +83,10 @@ class Audio(RaspIotResource):
             debug_enabled (bool): flag to set debug level to logger
         """
         #init
-        RaspIotResource.__init__(self, self.RESOURCES, bootstrap, debug_enabled)
+        RaspIotResources.__init__(self, bootstrap, debug_enabled)
 
         #members
-        self.alsa = Alsa(self.cleep_filesystem)
+        self.alsa = Alsa(self.drivers, self.cleep_filesystem)
         self.asoundrc = Asoundrc(self.cleep_filesystem)
 
     def _configure(self):
@@ -65,6 +96,10 @@ class Audio(RaspIotResource):
         if not self.asoundrc.exists():
             self.logger.debug(u'Create default audio configuration file')
             self.asoundrc.set_default_device(self.DEFAULT_DEVICE[u'card'], self.DEFAULT_DEVICE[u'device'])
+
+        #register default audio drivers (for jack and HDMI outputs)
+        for driver_name, driver in self.AUDIO_DRIVERS.items():
+            self._register_driver(AudioDriver(driver_name, driver))
 
     def get_module_config(self):
         """
@@ -86,13 +121,39 @@ class Audio(RaspIotResource):
         volumes = self.alsa.get_volumes()
 
         #improve configuration content
-        card_name = None
-        if current_config is not None:
-            for name in playback_devices.keys():
-                if playback_devices[name][u'cardid']==current_config[u'cardid']:
-                    card_name = name
-                    break
-            current_config[u'cardname'] = card_name
+        #card_name = None
+        #if current_config is not None:
+        #    for name in playback_devices.keys():
+        #        if playback_devices[name][u'cardid']==current_config[u'cardid']:
+        #            card_name = name
+        #            break
+        #    current_config[u'cardname'] = card_name
+        #search selected card according to current configuration
+        
+        
+
+        #improve audio devices using label
+        audio_resources = self._get_resources('audio\.')
+        if u'audio.playback' in audio_resources:
+            for device_name, device in playback_devices.items():
+                if device[u'cardid']==current_config[u'cardid'] and device[u'deviceid']==current_config[u'deviceid']:
+                    device.update({u'selected': True})
+                else:
+                    device.update({u'selected': False})
+                if device_name in audio_resources[u'audio.playback']:
+                    device.update({u'label': audio_resources[u'audio.playback'][device_name][u'label']})
+                else:
+                    device.update({u'label': u'<%s>' % device[u'name']})
+        if u'audio.capture' in audio_resources:
+            for device_name, device in capture_devices.items():
+                if device[u'cardid']==current_config[u'cardid'] and device[u'deviceid']==current_config[u'deviceid']:
+                    device.update({u'selected': True})
+                else:
+                    device.update({u'selected': False})
+                if device_name in audio_resources[u'audio.capture']:
+                    device.update({u'label': audio_resources[u'audio.capture'][device_name][u'label']})
+                else:
+                    device.update({u'label': u'<%s>' % device[u'name']})
 
         return {
             u'config': current_config,
@@ -148,61 +209,60 @@ class Audio(RaspIotResource):
         """
         Play test sound to make sure audio card is correctly configured
         """
-        #get playback resource
-        self.acquire_resource(u'audio.playback')
-
-        #play audio
-        if not self.alsa.play_sound(self.TEST_SOUND):
-            raise CommandError(u'Unable to play test sound: internal error.')
-
-        #release resource
-        self.release_resource(u'audio.playback')
+        #request playback resource
+        self._need_resource(u'audio.playback')
+        raise CommandInfo('Audio sample will be played in few seconds')
 
     def test_recording(self):
         """
         Record sound during few seconds and play it
         """
-        #get capture resource
-        self.acquire_resource(u'audio.capture')
+        #request capture resource
+        self._need_resource(u'audio.capture')
+        raise CommandInfo('Recording will start in few seconds. Please wait for message.')
 
-        #record sound
-        sound = self.alsa.record_sound(timeout=5.0)
-        self.logger.debug(u'Recorded sound: %s' % sound)
-        self.alsa.play_sound(sound)
-
-        #release resource
-        self.release_resource(u'audio.capture')
-
-        #purge file
-        time.sleep(0.5)
-        os.remove(sound)
-
-    def _acquire_resource(self, resource, extra):
+    def _resource_acquired(self, resource_name):
         """
-        Acquire resource
+        Function called when resource is acquired
 
         Args:
-            resource (string): resource name
-            extra (dict): extra parameters
-
-        Return:
-            bool: True if resource acquired
+            resource_name (string): acquired resource name
         """
-        #nothing to perform here
-        return True
+        self.logger.debug('Resource "%s" acquired' % resource_name)
+        if resource_name==u'audio.playback':
+            #play test sample
+            if not self.alsa.play_sound(self.TEST_SOUND):
+                raise CommandError(u'Unable to play test sound: internal error.')
 
-    def _release_resource(self, resource, extra):
+            #release resource
+            self._release_resource(u'audio.playback')
+
+        elif resource_name==u'audio.capture':
+            #record sound
+            sound = self.alsa.record_sound(timeout=5.0)
+            self.logger.debug(u'Recorded sound: %s' % sound)
+            self.alsa.play_sound(sound)
+
+            #release resource
+            self._release_resource(u'audio.capture')
+
+            #purge file
+            time.sleep(0.5)
+            try:
+                os.remove(sound)
+            except:
+                self.logger.warn(u'Unable to remove recorded test sound "%s"' % sound)
+
+        else:
+            self.logger.error(u'Unsupported resource "%s" acquired' % resource_name)
+
+    def _resource_needs_to_be_released(self, resource_name):
         """
-        Release resource
+        Function called when resource is acquired by other module and needs to be released.
 
         Args:
-            resource (string): resource name
-            extra (dict): extra parameters
-
-        Return:
-            bool: True if resource acquired
+            resource_name (string): acquired resource name
         """
-        #resource is not acquired during too much time, so do nothing
-        return True
-
+        #resource is not acquired for too long, it will be released naturally so nothing to do here
+        pass
